magnus-labrador-demo/requirements.txt
# Voice assistant dependencies (Mac dev + Raspberry Pi).
# Pinned loosely so pip can resolve platform-specific wheels (e.g. ARM vs x86).

# LLM client
openai>=2.0,<3

# STT — CTranslate2-based Whisper, runs on CPU with int8 quantization
faster-whisper>=1.0

# TTS — small ONNX models, optimized for Pi-class hardware
piper-tts>=1.2

# Audio I/O + DSP
sounddevice>=0.5
soundfile>=0.12
numpy>=1.26,<3
noisereduce>=3.0
